Complete redesign
This was a HUGE redesign from the ground up. I redid the code so it would be cleaner and passes code validation. I used a new apple inspired
UI design for a clean minimal approach. I also made sure to use a lot more styles and images. This will be the first of many big changes.